+++ javamelody-core/src/main/java/net/bull/javamelody/JavaMelodyLogger.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2008-2017 by Emeric Vernat
+ * Copyright 2008-2019 by Emeric Vernat
  *
  *     This file is part of Java Melody.
  *
@@ -72,5 +72,5 @@ public interface JavaMelodyLogger {
 	 * @param loggerName Nom du logger à utiliser
 	 */
 	void logHttpRequest(HttpServletRequest httpRequest, String requestName, long duration,
-			boolean systemError, int responseSize, String loggerName);
+			boolean systemError, long responseSize, String loggerName);
 }
