+++ javamelody-core/src/main/java/net/bull/javamelody/internal/web/CounterServletResponseWrapper.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2008-2017 by Emeric Vernat
+ * Copyright 2008-2019 by Emeric Vernat
  *
  *     This file is part of Java Melody.
  *
@@ -39,9 +39,9 @@ public class CounterServletResponseWrapper extends FilterServletResponseWrapper 
 
 	/**
 	 * Retourne la taille en octets du flux écrit dans la réponse.
-	 * @return int
+	 * @return long
 	 */
-	public int getDataLength() {
+	public long getDataLength() {
 		return getCounterResponseStream() == null ? 0 : getCounterResponseStream().getDataLength();
 	}
